# services/auth/cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for GoDelivery Auth Service.",
        "title": "GoDelivery Auth Service",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/auth/logout": {
            "delete": {
                "description": "Logout",
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "responses": {
                    "200": {
                        "description": "Logout success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "access token not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "description": "Refresh",
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh",
                "parameters": [
                    {
                        "description": "Refresh request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshReqBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Refresh success",
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshResp200Body"
                        }
                    },
                    "401": {
                        "description": "invalid credentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/auth/signin": {
            "post": {
                "description": "SignIn",
                "tags": [
                    "Auth"
                ],
                "summary": "SignIn",
                "parameters": [
                    {
                        "description": "SignIn request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.SignInReqBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SingIn success",
                        "schema": {
                            "$ref": "#/definitions/auth.SignInResp200Body"
                        }
                    },
                    "401": {
                        "description": "invalid credentials",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/auth/signup": {
            "post": {
                "description": "SingUp",
                "tags": [
                    "Auth"
                ],
                "summary": "SingUp",
                "parameters": [
                    {
                        "description": "SingUp request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.SignUpReqBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SingUp success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "user already exists",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/healthz": {
            "get": {
                "description": "Healthz check",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Healthz"
                ],
                "summary": "Healthz check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/healthz.HealthzRespBody"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.RefreshReqBody": {
            "type": "object",
            "required": [
                "refreshToken"
            ],
            "properties": {
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "auth.RefreshResp200Body": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "auth.SignInReqBody": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 12
                }
            }
        },
        "auth.SignInResp200Body": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "auth.SignUpReqBody": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 12
                }
            }
        },
        "healthz.HealthzRespBody": {
            "type": "object",
            "required": [
                "env",
                "status"
            ],
            "properties": {
                "env": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}